+++ auth/apps.py
@@ -2,6 +2,5 @@ from django.apps import AppConfig
 
 
 class AuthConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
     name = 'auth'
     label = 'authentication'
