+++ auth/views.py
@@ -1,36 +1,7 @@
 from rest_framework.permissions import AllowAny
 from .serializers import LoginSerializer
-from user.serializers import UserSerializer
-from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
-from rest_framework import viewsets
-from user.models import User
-from rest_framework.response import Response
-from rest_framework import status
+from rest_framework_simplejwt.views import TokenObtainPairView
 
 class LoginViewSet(TokenObtainPairView):
     permission_classes = [AllowAny]
     serializer_class = LoginSerializer
-
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
-
-        try:
-            serializer.is_valid(raise_exception=True)
-        except TokenError as e:
-            raise InvalidToken(e.args[0])
-
-        return Response(serializer.validated_data, status=status.HTTP_200_OK)
-
-class RegisterViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all()
-    permission_classes = [AllowAny]
-    serializer_class = UserSerializer
-
-    def create(self, request, *args, **kwargs):
-        permission_classes = [AllowAny]
-        serializer = self.get_serializer(data=request.data)
-        
-        serializer.is_valid(raise_exception=True)
-        serializer.save(created_by=self.request.user)
-        
-        return Response(status=status.HTTP_201_CREATED)
