+++ back/urls.py
@@ -11,14 +11,13 @@ from rest_framework_simplejwt.views import (
     TokenVerifyView,
 )
 
-from auth.views import LoginViewSet, RegisterViewSet
+from auth.views import LoginViewSet
 
 urlpatterns = [
 
     path('admin/', admin.site.urls),
 
     path('login/', LoginViewSet.as_view(), name='login_viewset'),
-    path('register/', RegisterViewSet.as_view({'post': 'create'}), name='register_viewset'),
     path('refresh/', TokenRefreshView.as_view(), name='token_refresh'),
     path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),
 
@@ -26,7 +25,4 @@ urlpatterns = [
     path('',include('movie.urls')),
     path('',include('genre.urls')),
     
-]
-
-if settings.DEBUG:
-    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+]+ static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
