+++ genre/apps.py
@@ -2,5 +2,4 @@ from django.apps import AppConfig
 
 
 class GenreConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
     name = 'genre'
