+++ movie/apps.py
@@ -2,5 +2,4 @@ from django.apps import AppConfig
 
 
 class MovieConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
     name = 'movie'
